requirements.txt
kafka-python # Thư viện Kafka client cho Python
Faker        # Thư viện tạo dữ liệu giả lập